Some of the WOPL Changes

+++ src/main/java/com/soapboxrace/core/bo/TokenSessionBO.java
@@ -115,33 +115,43 @@ public class TokenSessionBO {
 
         if (email != null && !email.isEmpty() && password != null && !password.isEmpty()) {
             UserEntity userEntity = userDAO.findByEmail(email);
+
             if (userEntity != null) {
-                if (password.equals(userEntity.getPassword())) {
-                    if (userEntity.isLocked()) {
-                        loginStatusVO.setDescription("Account locked. Contact an administrator.");
+                if(userEntity.isAdmin() || parameterBO.getBoolParam("IS_MAINTENANCE") == false) {
+                    if (password.equals(userEntity.getPassword())) {
+                        if (userEntity.isLocked()) {
+                            loginStatusVO.setDescription("Account locked. Contact an administrator.");
+                            return loginStatusVO;
+                        }
+
+                        BanEntity banEntity = authenticationBO.checkUserBan(userEntity);
+
+                        if (banEntity != null) {
+                            LoginStatusVO.Ban ban = new LoginStatusVO.Ban();
+                            ban.setReason(banEntity.getReason());
+                            if (banEntity.getEndsAt() != null)
+                                ban.setExpires(DateTimeFormatter.ofLocalizedDateTime(FormatStyle.FULL).withZone(ZoneId.systemDefault()).format(banEntity.getEndsAt()));
+                            loginStatusVO.setBan(ban);
+                            return loginStatusVO;
+                        }
+
+                        userEntity.setLastLogin(LocalDateTime.now());
+                        userDAO.update(userEntity);
+                        Long userId = userEntity.getId();
+                        deleteByUserId(userId);
+                        String randomUUID = createToken(userId, httpRequest.getRemoteHost());
+                        loginStatusVO = new LoginStatusVO(userId, randomUUID, true);
+                        loginStatusVO.setDescription("");
+
                         return loginStatusVO;
                     }
+                } else {
+                    String message = (parameterBO.getStrParam("IS_MAINTENANCE_MSG").isEmpty()) ? 
+                        parameterBO.getStrParam("IS_MAINTENANCE_MSG") : 
+                        "Server is in maintenance. Please follow our discord for more info.";
 
-                    BanEntity banEntity = authenticationBO.checkUserBan(userEntity);
-
-                    if (banEntity != null) {
-                        LoginStatusVO.Ban ban = new LoginStatusVO.Ban();
-                        ban.setReason(banEntity.getReason());
-                        if (banEntity.getEndsAt() != null)
-                            ban.setExpires(DateTimeFormatter.ofLocalizedDateTime(FormatStyle.FULL).withZone(ZoneId.systemDefault()).format(banEntity.getEndsAt()));
-                        loginStatusVO.setBan(ban);
-                        return loginStatusVO;
-                    }
-
-                    userEntity.setLastLogin(LocalDateTime.now());
-                    userDAO.update(userEntity);
-                    Long userId = userEntity.getId();
-                    deleteByUserId(userId);
-                    String randomUUID = createToken(userId, httpRequest.getRemoteHost());
-                    loginStatusVO = new LoginStatusVO(userId, randomUUID, true);
-                    loginStatusVO.setDescription("");
-
-                    return loginStatusVO;
+                    loginStatusVO.setDescription(message);
+                    return loginStatusVO;   
                 }
             }
         }
